+++ tests/e2e.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 
 # test if in docker environment
 # this script should only be running in docker!
@@ -10,8 +11,8 @@ else
   exit 1
 fi
 
-# start in tests/ dir
-cd "$(dirname "$0")"
+# start in the directory where the script runs
+cd $SCRIPT_DIR
 
 TEMP_APP_PATH=`mktemp -d`
 VERDACCIO_REGISTRY_URL=http://localhost:4873
@@ -49,6 +50,9 @@ yarn config set registry "$VERDACCIO_REGISTRY_URL"
 # login to custom registy
 (cd && npx npm-auth-to-token@1.0.0 -u user -p password -e user@example.com -r "$VERDACCIO_REGISTRY_URL")
 
+#==============================#
+# Test development environment #
+#==============================#
 # lint project
 yarn lint
 # test templates
@@ -57,11 +61,14 @@ yarn build
 exists template/dist/*.js
 CI=true yarn test
 
-# bump version
+# bump version so that we can 'deploy' to verdaccio
 npm version patch --force
 # publish to verdaccio
 npm publish
 
+#================================#
+# Test end user install from npm #
+#================================#
 # simulate end user installs
 cd $TEMP_APP_PATH
 npx typescript-node-scripts create test-app
@@ -69,10 +76,26 @@ npx typescript-node-scripts create test-app
 cd test-app
 exists node_modules/typescript-node-scripts
 
+#========================#
+# Test generated project #
+#========================#
 # run tslint on generated project
 yarn lint
 # build
 yarn build
 # check for build files
 exists dist/*.js
+CI=true yarn test
+
+#=================================================#
+# Run sink/ts-js-integration in generated project #
+#=================================================#
+# remove src folder
+rm -rf src/*
+# copy the ts-js-integration files to src
+cp -a "$SCRIPT_DIR/sink/ts-js-integration/." ./src
+# run the build and execute the bundle
+yarn build
+node dist/bundle.prod.js
+# run the test files
 CI=true yarn test