+++ tests/sink/ts-js-integration/esm.js
@@ -1,0 +1,8 @@
+import fs from 'fs'
+
+export const readFileSync = fs.readFileSync
+export const TEST = 2
+export default () => {
+    // tslint:disable-next-line
+    console.log('App is running')
+}