+++ tests/sink/ts-js-integration/ts.ts
@@ -1,0 +1,5 @@
+export const TEST = 2
+export default () => {
+    // tslint:disable-next-line
+    console.log('App is running')
+}